pull out public class to own file
